move ComponentId to separate file
